Netflix Share Market dashboard (Power BI)
In this project, we have extraction the share market data of the Netflix from the stock exchange and then create the excel sheet. After uploading the data in the Power Query we have done modelling and some changes in the data and with help of Power BI. I have created the dashboard I hope you will like it.